+++ src/App.js
@@ -1,39 +1,18 @@
-import CategoriesContainer from './Components/Container/CategoriesContainer';
+import { Route, Routes } from 'react-router-dom';
+import HomePage from './Components/Routes/HomePage/HomePage';
+import Navigation from './Components/Routes/Navigation/Navigation';
+import ShopPage from './Components/Routes/ShopPage/ShopPage';
 
 
 const App = () => {
-
-  const categories = [
-    {
-      "id": 1,
-      "title": "hats",
-      "imageUrl": "https://i.ibb.co/cvpntL1/hats.png"
-    },
-    {
-      "id": 2,
-      "title": "jackets",
-      "imageUrl": "https://i.ibb.co/px2tCc3/jackets.png"
-    },
-    {
-      "id": 3,
-      "title": "sneakers",
-      "imageUrl": "https://i.ibb.co/0jqHpnp/sneakers.png"
-    },
-    {
-      "id": 4,
-      "title": "womens",
-      "imageUrl": "https://i.ibb.co/GCCdy8t/womens.png"
-    },
-    {
-      "id": 5,
-      "title": "mens",
-      "imageUrl": "https://i.ibb.co/R70vBrQ/men.png"
-    }
-  ]
-
   return (
-    <CategoriesContainer categories={categories} />
+    <Routes>
+      <Route path='/' element={<Navigation />}>
+        <Route index element={<HomePage />} />
+        <Route path='shop' element={<ShopPage />} />
+      </Route>
+    </Routes>
   );
-}
+};
 
 export default App;
