+++ src/Components/Routes/ShopPage/ShopPage.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const ShopPage = () => {
+	return (
+		<div>ShopPage</div>
+	)
+}
+
+export default ShopPage
